requirements.txt
numpy
pandas
pillow
ffcv
pydicom
torch
torchvision
tqdm
